refactor(monoio): split h1 connection flow into modules
