+++ harrow-server-monoio/src/lib.rs
@@ -28,7 +28,7 @@
 //!           ▼                       ▼
 //! ┌─────────────────┐   ┌─────────────────────┐
 //! │   H1 Handler    │   │   H2 Handler        │
-//! │    (h1.rs)      │   │    (h2.rs)          │
+//! │(h1/dispatcher.rs)│  │    (h2.rs)          │
 //! └─────────────────┘   └─────────────────────┘
 //! ```
 //!
